refactor(utils): reorganize swarm utils into a module
